Tach.Impulse.Types.Impulse

Documentation

data ImpulseSeries key isp st en rep

'ImpulseSeries is the fundamental datatype that represents all the various kinds of time related event data.

It could be a-periodic sampling from a cell modem, a set of compressed files etc.. .

The idea is any set of values represented by 'rep is expected to have period 'isp, which may be a parameterized value. The series start value ('st) and end value ('en) are there to be used to identify individual data points.

|

Constructors

ImpulseSeries

Fields

impulseSeriesKey :: key

impulseSeriesPeriod :: isp

impulseSeriesStart :: st

impulseSeriesEnd :: en

impulseSeriesRep :: rep

newtype ImpulseKey a

some standard newtype helpers for defining isp, rep st en

Constructors

ImpulseKey

Fields

unKey :: a

newtype ImpulseStart a

Constructors

ImpulseStart

Fields

unStart :: a

newtype ImpulseEnd a

Constructors

ImpulseEnd

Fields

unEnd :: a

newtype ImpulseRep a

Constructors

ImpulseRep

Fields

unRep :: a

data ImpulsePeriod a b

Constructors

IPeriodConst b

IPeriodParameterized a