Tach.Impulse.Types.TimeValueSeries

Documentation

type TVSTypeOfTime = TVTypeOfTime

Time Value Series are based on the impulse series, and designed to be in the time domain. They may be based on a fixed period or a parameterized one. |

The kind of time used for the time value series

type TVSPeriod = TVPeriod

type TVSKey = TVKey

type TVSStart = ImpulseStart TVSTypeOfTime

type TVSEnd = ImpulseEnd TVSTypeOfTime

type TimeValueSeries rep = ImpulseSeries TVSKey TVSPeriod TVSStart TVSEnd (ImpulseRep rep)

TimeValueSeries are the generic representations of a series of values in the time domain. What is meant by 'time domain' and value are left to the specific implementation

type TVSDoublePeriodicType = TimeValueSeries (Vector Double)

TimeValue Standard Type is Double and this is the periodic version

type TVSimpleImpulseType = TimeValueSeries (Set TVNoKey)

TimeValue Standard Type is TVSimple and this is the impulse version